requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
absl-py=0.11.0=pypi_0
aiofiles=0.6.0=pypi_0
appdirs=1.4.4=pypi_0
astor=0.8.1=pypi_0
audioread=2.1.9=pypi_0
blas=1.0=mkl
blis=0.7.4=pypi_0
boto3=1.17.7=pypi_0
botocore=1.20.7=pypi_0
ca-certificates=2020.10.14=0
cached-property=1.5.2=pypi_0
cachetools=4.2.1=pypi_0
catalogue=2.0.1=pypi_0
certifi=2020.6.20=py37_0
cffi=1.14.5=pypi_0
chardet=4.0.0=pypi_0
click=7.1.2=py_0
cycler=0.10.0=pypi_0
cymem=2.0.5=pypi_0
datefinder=0.7.1=pypi_0
decorator=4.4.2=pypi_0
dill=0.3.3=pypi_0
en-core-web-sm=3.0.0=pypi_0
fastapi=0.63.0=pypi_0
filelock=3.0.12=pypi_0
flask=1.1.2=py_0
freetype=2.10.4=ha233b18_0
gast=0.2.2=pypi_0
google-api-core=1.26.0=pypi_0
google-auth=1.26.1=pypi_0
google-cloud-core=1.6.0=pypi_0
google-cloud-speech=2.0.1=pypi_0
google-cloud-storage=1.36.0=pypi_0
google-crc32c=1.1.2=pypi_0
google-pasta=0.2.0=pypi_0
google-resumable-media=1.2.0=pypi_0
googleapis-common-protos=1.52.0=pypi_0
grpcio=1.35.0=pypi_0
h11=0.12.0=pypi_0
h5py=3.1.0=pypi_0
idna=2.10=pypi_0
imageio=2.9.0=pypi_0
imageio-ffmpeg=0.4.3=pypi_0
importlib-metadata=3.4.0=pypi_0
inflect=5.0.2=pypi_0
intel-openmp=2019.4=233
itsdangerous=1.1.0=py37_0
jinja2=2.11.2=py_0
jmespath=0.10.0=pypi_0
joblib=1.0.1=pypi_0
jpeg=9b=he5867d9_2
jsonpatch=1.28=pypi_0
jsonpointer=2.0=pypi_0
keras-applications=1.0.8=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.3.1=pypi_0
lcms2=2.11=h92f6f08_0
libcst=0.3.17=pypi_0
libcxx=10.0.0=1
libedit=3.1.20191231=h1de35cc_1
libffi=3.3=hb1e8313_2
libpng=1.6.37=ha441bb4_0
librosa=0.8.0=pypi_0
libtiff=4.1.0=hcb84e12_1
libuv=1.40.0=haf1e3a3_0
llvmlite=0.35.0=pypi_0
lz4-c=1.9.3=h23ab428_0
markdown=3.3.3=pypi_0
markupsafe=1.1.1=py37h1de35cc_0
matplotlib=3.3.4=pypi_0
mkl=2019.4=233
mkl-service=2.3.0=py37h9ed2024_0
mkl_fft=1.2.0=py37hc64f4ea_0
mkl_random=1.1.1=py37h959d312_0
moviepy=1.0.3=pypi_0
multiprocess=0.70.11.1=pypi_0
murmurhash=1.0.5=pypi_0
mypy-extensions=0.4.3=pypi_0
ncurses=6.2=h0a44026_1
ninja=1.10.2=py37hf7b0b51_0
nltk=3.5=pypi_0
numba=0.52.0=pypi_0
numpy=1.19.4=pypi_0
olefile=0.46=py37_0
openai=0.3.0=pypi_0
openssl=1.1.1h=haf1e3a3_0
opt-einsum=3.3.0=pypi_0
packaging=20.9=pypi_0
pathy=0.3.6=pypi_0
pillow=8.1.0=py37h5270095_0
pip=20.3.3=py37hecd8cb5_0
pooch=1.3.0=pypi_0
preshed=3.0.5=pypi_0
proglog=0.1.9=pypi_0
proto-plus=1.13.0=pypi_0
protobuf=3.14.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.20=pypi_0
pydantic=1.7.3=pypi_0
pydub=0.24.1=pypi_0
pynndescent=0.5.2=pypi_0
pyparsing=2.4.7=pypi_0
pyqt5=5.15.2=pypi_0
pyqt5-sip=12.8.1=pypi_0
python=3.7.9=h26836e1_0
python-dateutil=2.8.1=pypi_0
pytorch=1.7.1=py3.7_0
pytube3=9.6.4=pypi_0
pytz=2021.1=pypi_0
pyyaml=5.4.1=pypi_0
pyzmq=22.0.3=pypi_0
readline=8.1=h9ed2024_0
regex=2020.11.13=pypi_0
requests=2.25.1=pypi_0
resampy=0.2.2=pypi_0
rsa=4.7=pypi_0
s3transfer=0.3.4=pypi_0
sacremoses=0.0.43=pypi_0
scikit-learn=0.24.1=pypi_0
scipy=1.6.0=pypi_0
sense2vec=2.0.0=pypi_0
sentencepiece=0.1.95=pypi_0
setuptools=52.0.0=py37hecd8cb5_0
six=1.15.0=py37hecd8cb5_0
smart-open=3.0.0=pypi_0
sounddevice=0.4.1=pypi_0
soundfile=0.10.3.post1=pypi_0
spacy=3.0.1=pypi_0
spacy-legacy=3.0.1=pypi_0
sqlite=3.33.0=hffcf06c_0
srsly=2.4.0=pypi_0
starlette=0.13.6=pypi_0
tensorboard=1.15.0=pypi_0
tensorflow=1.15.0=pypi_0
tensorflow-estimator=1.15.1=pypi_0
termcolor=1.1.0=pypi_0
text2text=0.0.9=pypi_0
thinc=8.0.1=pypi_0
threadpoolctl=2.1.0=pypi_0
tk=8.6.10=hb0a8c7a_0
tokenizers=0.10.1=pypi_0
torchaudio=0.7.2=py37
torchfile=0.1.0=pypi_0
torchvision=0.8.2=py37_cpu
tornado=6.1=pypi_0
tqdm=4.56.2=pypi_0
transformers=4.3.2=pypi_0
typer=0.3.2=pypi_0
typing-inspect=0.6.0=pypi_0
typing_extensions=3.7.4.3=pyha847dfd_0
umap-learn=0.5.1=pypi_0
unidecode=1.2.0=pypi_0
urllib3=1.26.3=pypi_0
uvicorn=0.13.3=pypi_0
visdom=0.1.8.9=pypi_0
wasabi=0.8.2=pypi_0
webrtcvad=2.0.10=pypi_0
websocket-client=0.57.0=pypi_0
werkzeug=1.0.1=py_0
wheel=0.36.2=pyhd3eb1b0_0
wrapt=1.12.1=pypi_0
xz=5.2.5=h1de35cc_0
zipp=3.4.0=pypi_0
zlib=1.2.11=h1de35cc_3
zstd=1.4.5=h41d2c2f_0
